refactor: extract progress.rs from backend/mod.rs
Reduces backend/mod.rs from 1288 to 1224 lines.
Part of upstream codebase refactoring (commit 98d215d).
914 tests pass.
